Separate add button handler from munu button hander
