+++ src/index.js
@@ -1,4 +1,5 @@
-import { menuButtonHandler, addButtonHandler } from "./menu-button-handler.js";
+import menuButtonHandler from "./menu-button-handler.js";
+import addButtonHandler from "./add-button-handler.js";
 
 document.addEventListener("click", menuButtonHandler);
 document.addEventListener("click", addButtonHandler);
