+++ src/menu-button-handler.js
@@ -1,4 +1,4 @@
-import StationManager from "./station-manager.js";
+// import StationManager from "./station-manager.js";
 import { getChildById } from "./custom-dom-handler.js";
 
 const MENU_BUTTONS_ID = {
@@ -6,18 +6,6 @@ const MENU_BUTTONS_ID = {
   lineManagerButton: "line-manager-button",
   sectionManagerButton: "section-manager-button",
   mapPrintManagerButton: "map-print-manager-button",
-};
-
-const ADD_BUTTONS_ID = {
-  stationAddButton: "station-add-button",
-  lineAddButton: "line-add-button",
-  sectionAddButton: "section-add-button",
-};
-
-const INPUT_FORM_ID = {
-  stationNameInput: "station-name-input",
-  lineNameInput: "line-name-input",
-  sectionOrderInput: "section-order-input",
 };
 
 const renderStationAddForm = (container) => {
@@ -41,46 +29,8 @@ const showStationManagerPage = (e) => {
   renderStationAddForm($mainContentsContainer);
 };
 
-const isInputStringLengthTwoOrGreater = (input) => {
-  if (input.length < 2) {
-    alert("두 글자 이상 입력해주세요.");
-    return false;
-  }
-  return true;
-};
-
-const isStationNameInputDuplicate = (input) => {
-  if (new StationManager().hasName(input)) {
-    alert("해당 역이 이미 존재합니다.");
-    return true;
-  }
-  return false;
-};
-
-const isValidStationName = (name) => {
-  return (
-    isInputStringLengthTwoOrGreater(name) && !isStationNameInputDuplicate(name)
-  );
-};
-
-const stationAddButtonHandler = (e) => {
-  const stationNameInputValue = getChildById(
-    e.target.parentNode.children[0],
-    INPUT_FORM_ID.stationNameInput
-  ).value;
-  if (isValidStationName(stationNameInputValue)) {
-    new StationManager().addStation(stationNameInputValue);
-  }
-};
-
-export const menuButtonHandler = (e) => {
+export default function menuButtonHandler(e) {
   if (e.target.id === MENU_BUTTONS_ID.stationManagerButton) {
     showStationManagerPage(e);
   }
-};
-
-export const addButtonHandler = (e) => {
-  if (e.target.id === ADD_BUTTONS_ID.stationAddButton) {
-    stationAddButtonHandler(e);
-  }
-};
+}
